Refactor: 가격 long, 수량 BigDecimal로 전환 (KafakaOrderProducer 포함 시그니처 변경)
- KRW 가격 표현을 double → long(정수)로 변경하여 부동소수점 오차 제거 및 일관성 확보
- 수량/잔량/누적체결량을 double → BigDecimal로 전환, 검증/계산 로직 및 이벤트 페이로드 반영
- KafkaOrderProducer 메서드 시그니처와 호출부 전면 수정(vwap/last/limit=long, filled/remaining=BigDecimal)

+++ matching-engine/src/main/java/com/beyond/MKX/infrastructure/kafka/KafkaOrderProducer.java
@@ -7,6 +7,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -34,7 +35,7 @@ public class KafkaOrderProducer {
     // 체결 이벤트 (부분/완전 체결 1건마다 발행)
     // ---------------------------------------------------------------------
     public void sendExecution(String marketOrderId, String ticker, String side,
-                              String counterOrderId, double qty, double price) {
+                              String counterOrderId, BigDecimal qty, long price) {
         String execId = marketOrderId + "-" + counterOrderId + "-" + UUID.randomUUID(); // 멱등키
         ExecutionEvent evt = ExecutionEvent.builder()
                 .execId(execId)
@@ -64,7 +65,7 @@ public class KafkaOrderProducer {
     // 상태 알림 (접수/대기/부분/완전/취소)
     // ---------------------------------------------------------------------
     /** 신규 지정가 접수(체결 전): price=지정가, remaining=초기 수량 */
-    public void sendNewAccepted(String orderId, String ticker, String side, double price, double qty) {
+    public void sendNewAccepted(String orderId, String ticker, String side, long price, BigDecimal qty) {
         OrderStatusEvent evt = OrderStatusEvent.builder()
                 .orderId(orderId)
                 .ticker(ticker)
@@ -99,8 +100,8 @@ public class KafkaOrderProducer {
 
     /** 부분 체결(대표가격=vwap, 보조=last/limit, 누적 체결수량 포함) */
     public void sendMarketPartial(String orderId, String ticker, String side,
-                                  double remaining,
-                                  double vwap, double lastPrice, double limitPrice, double filledQty) {
+                                  BigDecimal remaining,
+                                  long vwap, long lastPrice, long limitPrice, BigDecimal filledQty) {
         OrderStatusEvent evt = OrderStatusEvent.builder()
                 .orderId(orderId)
                 .ticker(ticker)
@@ -118,7 +119,7 @@ public class KafkaOrderProducer {
 
     /** 완전 체결(대표가격=vwap, last/limit/누적 체결수량 포함) */
     public void sendMarketFilled(String orderId, String ticker, String side,
-                                 double vwap, double lastPrice, double limitPrice, double filledQty) {
+                                 long vwap, long lastPrice, long limitPrice, BigDecimal filledQty) {
         OrderStatusEvent evt = OrderStatusEvent.builder()
                 .orderId(orderId)
                 .ticker(ticker)
